backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.40.0
httpx==0.27.0
aiosqlite==0.20.0
python-dotenv==1.0.1
websockets==13.0
pydantic==2.9.0
apscheduler==3.10.4
